+++ lib/sim/debug/expr_grammar.txt
@@ -8,5 +8,5 @@ p2       = p1 ([*/%] p2)?
 p1       = [*+-~!&]? p0
 p0       = value | IDENT ( . | ->) IDENT
 value    = register | constant | # IDENT | IDENT | ( p7 )
-constant = [0-9]+ | 0[xX][0-9a-fA-F]+ | 0[bB][01]+
+constant = [0-9]+ | 0[xX][0-9a-fA-F]+ | 0[bB][01]+ (_[iu](8|16|32))
 register = $ IDENT
